[vpr][route][crr] remove double from acceptable values for cell

+++ vpr/src/route/rr_graph_generation/tileable_rr_graph/crr_generator/data_frame_processor.cpp
@@ -299,7 +299,7 @@ Cell DataFrameProcessor::parse_csv_cell(const std::string& value) {
     // Try to parse as number
     try {
         size_t pos;
-        double num = std::stod(trimmed, &pos);
+        int num = std::stoi(trimmed, &pos);
         if (pos == trimmed.length()) {
             return Cell(num);
         }
